fix don`t view inactive EditText border and caption fix the new check replays check with id =0 into database

+++ app/src/main/java/com/friendroids/moneymana/db/PurchaseDB.kt
@@ -159,7 +159,7 @@ class PurchaseDB(applicationContext: Context) {
 
     private fun toCheckEntity(check: Check) = CheckEntity(
             id = check.id,
-            idCategory = check.category.id.toLong(),
+            idCategory = check.category.id,
             dateCheck = check.dateCheck,
             sumCheck = check.sumCheck,
             fnCheck = check.fnCheck,
